+++ pkg/user/repository/userRepository.go
@@ -5,7 +5,7 @@ import "github.com/jaayroots/todo-api/entities"
 type UserRepository interface {
 	Create(user *entities.User) (*entities.User, error)
 	FindByEmail(email string) (*entities.User, error)
-	GetByUserID(userID uint64) (*entities.User, error)
-	Update(userID uint64, user *entities.User) (*entities.User, error)
-	Delete(userID uint64) error
+	GetByUserID(userID int) (*entities.User, error)
+	Update(userID int, user *entities.User) (*entities.User, error)
+	Delete(userID int) error
 }
